Modernize kernel autoload bootstrap
