+++ phpunit.xml.dist
@@ -9,8 +9,8 @@
 >
     <php>
         <ini name="error_reporting" value="-1" />
-        <server name="KERNEL_DIR" value="app/" />
-        <server name="KERNEL_CLASS" value="AppKernel" />
+        <server name="KERNEL_DIR" value="src/" />
+        <server name="KERNEL_CLASS" value="App\Kernel" />
         <env name="SYMFONY_DEPRECATIONS_HELPER" value="weak" />
         <env name="SYMFONY_PHPUNIT_VERSION" value="9.6" />
     </php>
