+++ src/Kernel.php
@@ -1,0 +1,60 @@
+<?php
+
+namespace App;
+
+use Bazinga\Bundle\HateoasBundle\BazingaHateoasBundle;
+use Devlabs\SportifyBundle\DevlabsSportifyBundle;
+use Doctrine\Bundle\DoctrineBundle\DoctrineBundle;
+use JMS\SerializerBundle\JMSSerializerBundle;
+use Symfony\Bundle\DebugBundle\DebugBundle;
+use Symfony\Bundle\FrameworkBundle\FrameworkBundle;
+use Symfony\Bundle\MonologBundle\MonologBundle;
+use Symfony\Bundle\SecurityBundle\SecurityBundle;
+use Symfony\Bundle\TwigBundle\TwigBundle;
+use Symfony\Bundle\WebProfilerBundle\WebProfilerBundle;
+use Symfony\Component\Config\Loader\LoaderInterface;
+use Symfony\Component\HttpKernel\Kernel as BaseKernel;
+
+class Kernel extends BaseKernel
+{
+    public function registerBundles(): iterable
+    {
+        $bundles = [
+            new FrameworkBundle(),
+            new SecurityBundle(),
+            new TwigBundle(),
+            new MonologBundle(),
+            new DoctrineBundle(),
+            new DevlabsSportifyBundle(),
+            new JMSSerializerBundle(),
+            new BazingaHateoasBundle(),
+        ];
+
+        if (in_array($this->getEnvironment(), ['dev', 'test'], true)) {
+            $bundles[] = new DebugBundle();
+            $bundles[] = new WebProfilerBundle();
+        }
+
+        return $bundles;
+    }
+
+    public function getRootDir(): string
+    {
+        return $this->getProjectDir().'/app';
+    }
+
+    public function getCacheDir(): string
+    {
+        return $this->getProjectDir().'/var/cache/'.$this->getEnvironment();
+    }
+
+    public function getLogDir(): string
+    {
+        return $this->getProjectDir().'/var/logs';
+    }
+
+    public function registerContainerConfiguration(LoaderInterface $loader): void
+    {
+        $loader->load($this->getRootDir().'/config/config_'.$this->getEnvironment().'.yml');
+    }
+}
