+++ web/app.php
@@ -1,5 +1,6 @@
 <?php
 
+use App\Kernel;
 use Symfony\Component\HttpFoundation\Request;
 
 /**
@@ -9,7 +10,7 @@ $loader = require __DIR__.'/../app/autoload.php';
 
 define('WEB_DIRECTORY', __DIR__);
 
-$kernel = new AppKernel('prod', false);
+$kernel = new Kernel('prod', false);
 //$kernel = new AppCache($kernel);
 
 // When using the HttpCache, you need to call the method in your front controller instead of relying on the configuration parameter
